+++ HISTORY.rst
@@ -13,8 +13,7 @@ Pending Release
   ``importlib.metadata.version("flake8-tidy-imports")``
   (`docs <https://docs.python.org/3.8/library/importlib.metadata.html#distribution-versions>`__ /
   `backport <https://pypi.org/project/importlib-metadata/>`__).
-
-* Add dependencies on ``cached-property`` and ``importlib-metadata``.
+* Add dependency on ``importlib-metadata``.
 
 3.0.0 (2019-10-15)
 ------------------